Randomize stimuli over grid, make grid smaller
The stimuli were not properly randomized over the 3*3 grid, four of the
the position were never used. This is fixed.
Additionally, The entire scree was used, but that doesn't fit behind the
canvas, so the grid is also made a bit smaller.
Additionally the code is made a bit simpler.